# query/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server Classroom service",
        "title": "Classroom API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/c/classroom",
    "paths": {
...
        "/modules/download_file/{idFile}/{idModule}": {
            "get": {
                "description": "Download file of module",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules",
                    "classroom",
                    "roles.teacher",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Download file module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idFile",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.TokenMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad path params",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/modules/get_module/{idModule}": {
            "get": {
                "description": "Get module",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules",
                    "classroom",
                    "roles.teacher",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Get module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.ModuleMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/modules/get_modules": {
            "get": {
                "description": "Get user modules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules",
                    "classroom",
                    "roles.teacher",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Get modules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.ModulesMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "403": {
                        "description": "No estás asignado a ningún curso",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/modules/get_modules_history": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules",
                    "roles.student",
                    "roles.student_directive"
                ],
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Get total?",
                        "name": "total",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Skip",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad query param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/modules/search/{idModule}": {
            "get": {
                "description": "search in all module content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules",
                    "classroom",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desc",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.SearchHitsMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Attached": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "file": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "link": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "https://example.com"
                },
                "title": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "This is a title"
                },
                "type": {
                    "description": "Link or file",
                    "type": "string",
                    "enum": [
                        "link",
                        "file"
                    ],
                    "example": "link"
                }
            }
        },
        "models.Course": {
            "type": "object",
            "properties": {
                "__v": {
                    "type": "integer"
                },
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "course": {
                    "type": "string",
                    "example": "First"
                }
            }
        },
...
        "models.File": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "filename": {
                    "type": "string",
                    "example": "filename.png"
                },
                "key": {
                    "type": "string",
                    "example": "$53r34re"
                },
                "permissions": {
                    "type": "string",
                    "enum": [
                        "private",
                        "public",
                        "public_classroom"
                    ],
                    "example": "private"
                },
                "status": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string",
                    "example": "Title"
                },
                "type": {
                    "type": "string",
                    "example": "application/pdf"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/file"
                }
            }
        },
...
        "models.ModuleWithLookup": {
            "type": "object",
            "properties": {
                "__v": {
                    "type": "integer"
                },
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "section": {
                    "$ref": "#/definitions/models.Section"
                },
                "semester": {
                    "$ref": "#/definitions/models.Semester"
                },
                "status": {
                    "type": "boolean"
                },
                "sub_sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SubSection"
                    }
                },
                "subject": {
                    "$ref": "#/definitions/models.Subject"
                },
                "works": {
                    "description": "Works",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Work"
                    },
                    "x-omitempty": true
                }
            }
        },
        "models.Section": {
            "type": "object",
            "properties": {
                "__v": {
                    "type": "integer"
                },
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "course": {
                    "$ref": "#/definitions/models.Course"
                },
                "file": {
                    "$ref": "#/definitions/models.File"
                },
                "section": {
                    "type": "string",
                    "example": "A"
                }
            }
        },
        "models.Semester": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "semester": {
                    "type": "integer",
                    "example": 2
                },
                "year": {
                    "type": "integer",
                    "example": 2022
                }
            }
        },
...
        "models.SubSection": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "name": {
                    "type": "string",
                    "example": "Math - Geo"
                }
            }
        },
        "models.Subject": {
            "type": "object",
            "properties": {
                "__v": {
                    "type": "integer"
                },
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "subject": {
                    "type": "string",
                    "example": "Math"
                }
            }
        },
        "models.Work": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "acumulative": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "attached": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attached"
                    },
                    "x-omitempty": true
                },
                "author": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "date_limit": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "date_start": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "date_update": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "date_upload": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "description": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "This is a description"
                },
                "form": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "form_access": {
                    "type": "string",
                    "enum": [
                        "default",
                        "wtime"
                    ],
                    "x-omitempty": true,
                    "example": "default"
                },
                "grade": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "is_qualified": {
                    "type": "boolean"
                },
                "is_revised": {
                    "type": "boolean"
                },
                "module": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "pattern": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WorkPattern"
                    },
                    "x-omitempty": true
                },
                "time_access": {
                    "type": "integer",
                    "x-omitempty": true,
                    "example": 2
                },
                "title": {
                    "type": "string",
                    "example": "Work!"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "files",
                        "form"
                    ],
                    "example": "form"
                }
            }
        },
        "models.WorkPattern": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637d5de216f58bc8ec7f7f51"
                },
                "description": {
                    "type": "string",
                    "example": "This is a description"
                },
                "points": {
                    "type": "integer",
                    "example": 25
                },
                "title": {
                    "type": "string",
                    "example": "Pattern!"
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "smaps.ModuleMap": {
            "type": "object",
            "properties": {
                "module": {
                    "$ref": "#/definitions/models.ModuleWithLookup"
                }
            }
        },
        "smaps.ModulesMap": {
            "type": "object",
            "properties": {
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ModuleWithLookup"
                    }
                }
            }
        },
...
        "smaps.SearchHitsMap": {
            "type": "object",
            "properties": {
                "hits": {}
            }
        },
...
        "smaps.TokenMap": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}